% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affy.R
\name{group.to.median.sorted}
\alias{group.to.median.sorted}
\title{group to median}
\usage{
group.to.median.sorted(x, f)
}
\arguments{
\item{x}{data}

\item{f}{factor}
}
\description{
for preprocessing affymetrix SNP6 array
}

