% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_packages.R
\name{load_packages}
\alias{load_packages}
\title{Load packages (and install them, if need be)}
\usage{
load_packages(..., require = TRUE)
}
\arguments{
\item{...}{the names of packages as unquoted, comma-separated strings}

\item{require}{If TRUE (the default), the packages are loaded. If FALSE, the function only checks whether they are installed and installs them if not.}
}
\value{
inherited from require.
}
\description{
Check if a package is installed. If not, install it before
}
\examples{
\dontrun{
load_packages(package1, package2, package3)
}
}
