% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettydf.R
\name{pretty_df}
\alias{pretty_df}
\title{Pretty dataframes}
\usage{
pretty_df(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
a dataframe.
}
\description{
Convenience function for nice dataframe output
}
\examples{
df <- data.frame(a = c("a", "b", "c"),
                 b = c(0.86965965, 989687.89068, 278.9058574))
pretty_df(df)
}
