% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbarplot.R
\name{qbarplot}
\alias{qbarplot}
\title{Quick barplots}
\usage{
qbarplot(df, x, fill, wrap = NULL, raw = FALSE, text = TRUE,
  vjust_strip, cex = 3, cex_axes = 18, percent = TRUE,
  dodge = FALSE, xlab, ylab, wrap100 = FALSE, tibble = TRUE,
  display0 = FALSE, angle = 0, hjust = 0.5, vjust = 0.5, color,
  rev.palette = FALSE, ...)
}
\arguments{
\item{df}{a data frame with raw data}

\item{x}{variable to be displayed on the x axis. Must be a column name of df. Can be unquoted.}

\item{fill}{variable that defines the fill of the bars. Must be a column name of df. Can be unquoted.}

\item{wrap}{variable along which the data should be split in different panels. See ?ggplot2::facet_wrap and ?ggplot2::facet_grid for more information. Default is NULL, i.e. no wrap.}

\item{raw}{if TRUE, the function returns the raw data, i.e. a data frame with cross-tabulated absolute and relative frequencies.}

\item{text}{if TRUE (the default), absolute frequencies are displayed on the bars using geom_text.}

\item{vjust_strip}{vertical adjustment of the text, if text = TRUE. Behaves differently for stacked and dodged barplots. Under the hood, vjust is part of the position_stacked() call in stacked barplots (default: .5, i.e. in the middle of each portion of the bar) and part of the geom_text() call in dodged barplots (-.5, i.e. slightly above each bar).}

\item{cex}{size of the text displaying the absolute frequencies.}

\item{cex_axes}{size of the axis labels, 18pt by default. To get the ggplot2 default values and customize with your own ggplot layers, just enter a non-numeric value.}

\item{percent}{If TRUE (the default), the y axis displays proportions in percent format. Requires package "scales". If FALSE (or if "scales" is not installed), it is displayed as a fraction between 0 and 1.}

\item{dodge}{If FALSE (the default), a stacked barplot if produced. If TRUE, qbarplot() returns a dodged barplot, i.e. the bars are displayed side-by-side.}

\item{xlab}{x axis label. By default, identical to x.}

\item{ylab}{y axis label. By default, "Relative Frequency".}

\item{wrap100}{If wrap != NULL and wrap100 = TRUE, percentages are calculated separately for each category. In other words, if wrap100 = TRUE, the bars sum up to 100 percent in each panel. If wrap100 = FALSE (the default), the percentages sum up to 100 across panels. For illustration, compare e.g. \code{qbarplot(ggplot2::diamonds, color, clarity, cut, wrap100 = FALSE)} vs. \code{qbarplot(ggplot2::diamonds, color, clarity, cut, wrap100 = TRUE)}.}

\item{tibble}{If raw = TRUE and tibble = TRUE (the default), the raw data are returned as a tibble. Only works if the "tibble" package is installed. See ?tibble::tibble for more information.}

\item{display0}{If text = TRUE, display0 = TRUE (the default) prevents 0 values (if there are any) from being shown on the text display.}

\item{angle}{Use this to rotate the x tick labels.}

\item{hjust}{horizontal justification of x tick labels.}

\item{vjust}{vertical justification of x tick labels.}

\item{color}{color scheme to be used for the plot. Can be an ordinary palette (see ?palettes) or an RColorBrewer palette, see ?scale_fill_brewer for available palettes. Set to FALSE to get ggplot's default color scheme.}

\item{rev.palette}{if a color palette is used, setting rev.palette to TRUE will reverse it.}

\item{...}{Further arguments to be passed on to ggplot().}
}
\value{
a ggplot2 barplot or, if raw = TRUE, a data frame with the cross-tabulated raw and relative frequencies.
}
\description{
Convenience function for generating ggplot2 barplots from raw datasets.
}
\examples{
qbarplot(ggplot2::diamonds, color, cut)
qbarplot(ggplot2::diamonds, color, clarity, cut)
}
