% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbeeswarm.R
\name{qbeeswarm}
\alias{qbeeswarm}
\title{Quick beeswarm-boxplots}
\usage{
qbeeswarm(
  df,
  x,
  y,
  xlab,
  ylab,
  box = TRUE,
  notch = TRUE,
  fill_box = "lightblue",
  alpha = 0.3,
  cex_axes = 18,
  grid = FALSE,
  colour,
  bw = FALSE,
  ...
)
}
\arguments{
\item{df}{a data frame}

\item{x}{A column name within df, the variable to be displayed on the x axis.}

\item{y}{A column name within df, the variable to be displayed on the y axis.}

\item{xlab}{(optional) add x axis label manually.}

\item{ylab}{(optional) add y axis label manually.}

\item{box}{Should the boxplot be displayed? Default is TRUE.}

\item{notch}{Should the boxplot be displayed with a notch? Default is TRUE.}

\item{fill_box}{fill colour (or value) of the boxplot.}

\item{alpha}{alpha level of the boxplot color. Default is .3.}

\item{cex_axes}{font size of tick labels and axis text. Default is 18. Set to FALSE to add manual controls.}

\item{grid}{Should a grid be displayed? Default is FALSE.}

\item{colour}{colour of the points in the beeswarm plot}

\item{bw}{black-and-white: If set to TRUE, the plot will be printed in black and white.
Does not work properly if colour is specified. Overrides fill_box.}

\item{...}{further arguments to be passed on to ggplot.}
}
\value{
a ggplot2 beeswarm-and-boxplot.
}
\description{
convenience function for making beeswarmplots with overlaid boxplots.
}
\examples{
set.seed(200)
df <- data.frame(A = c(rep("high", 50), rep("lo", 50)),
                B = round(runif(n = 100)*100))
qbeeswarm(df, x = A, y = B)


}
