% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affy.R
\name{read.cel.signals}
\alias{read.cel.signals}
\title{read signal from one CEL file}
\usage{
read.cel.signals(cel.fn)
}
\arguments{
\item{cel.fn}{path to CEL}
}
\value{
SignalSet
}
\description{
read signal from one CEL file
}

