% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affy.R
\name{sset.cn.signal}
\alias{sset.cn.signal}
\title{get copy number signal from SignalSet}
\usage{
sset.cn.signal(sset)
}
\arguments{
\item{sset}{SignalSet}
}
\value{
sorted GRanges with signals
}
\description{
get copy number signal from SignalSet
}

