% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affy.R
\name{sset.snp.signal}
\alias{sset.snp.signal}
\title{get SNP signals from SignalSet}
\usage{
sset.snp.signal(sset)
}
\arguments{
\item{sset}{SignalSet}
}
\value{
GRanges with signals
}
\description{
get SNP signals from SignalSet
}

