% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_intensity.R
\name{visualizeGeneIntensity}
\alias{visualizeGeneIntensity}
\title{visualize gene}
\usage{
visualizeGeneIntensity(geneName, data, platform = "snp6", upstream = 2000,
  dwstream = 2000, refversion = "hg38", ...)
}
\arguments{
\item{geneName}{gene name}

\item{data}{data matrix (row: probes, column: samples)}

\item{platform}{snp6, EPIC, hm450}

\item{upstream}{distance to extend upstream}

\item{dwstream}{distance to extend downstream}

\item{refversion}{hg19 or hg38}

\item{...}{additional options, see visualizeRegionIntensity}
}
\description{
visualize gene
}

