% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_intensity.R
\name{visualizeProbesIntensity}
\alias{visualizeProbesIntensity}
\title{visualize probes}
\usage{
visualizeProbesIntensity(probeNames, data, platform = "EPIC",
  refversion = "hg38", upstream = 1000, dwstream = 1000, ...)
}
\arguments{
\item{probeNames}{probe names}

\item{data}{data matrix (row: probes, column: samples)}

\item{platform}{snp6}

\item{refversion}{hg19 or hg38}

\item{upstream}{distance to extend upstream}

\item{dwstream}{distance to extend downstream}

\item{...}{additional options, see visualizeRegionIntensity}
}
\description{
visualize probes
}

