% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_intensity.R
\name{visualizeRegionIntensity}
\alias{visualizeRegionIntensity}
\title{visualize region}
\usage{
visualizeRegionIntensity(chrm, plt.beg, plt.end, data, platform = "EPIC",
  refversion = "hg19", heat.height = NULL, draw = TRUE,
  show.sampleNames = TRUE, show.samples.n = NULL, show.probeNames = TRUE,
  na.rm = FALSE, dmin = 0, dmax = 1)
}
\arguments{
\item{chrm}{chromosome}

\item{plt.beg}{begin of the region}

\item{plt.end}{end of the region}

\item{data}{data matrix (row: probes, column: samples)}

\item{platform}{hm450 (default) or EPIC}

\item{refversion}{hg19}

\item{heat.height}{heatmap height (auto inferred based on rows)}

\item{draw}{draw figure or return data}

\item{show.sampleNames}{whether to show sample names}

\item{show.samples.n}{number of samples to show (default: all)}

\item{show.probeNames}{whether to show probe names}

\item{na.rm}{remove probes with all NA.}

\item{dmin}{data minimum on color scale}

\item{dmax}{data maximum on color scale}
}
\description{
visualize region
}

