% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_predictors.R
\name{wiz_add_predictors_group_modify}
\alias{wiz_add_predictors_group_modify}
\title{New add_predictors function using group_modify}
\usage{
wiz_add_predictors_group_modify(
  wiz_frame = NULL,
  variable = NULL,
  category = NULL,
  lookback = hours(48),
  window = lookback,
  stats = c(mean = mean, min = min, max = max),
  impute = TRUE,
  output_file = TRUE
)
}
\description{
New add_predictors function using group_modify
}
