% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_predictors.R
\name{wiz_add_predictors_old}
\alias{wiz_add_predictors_old}
\title{Old add predictors function using for() loops}
\usage{
wiz_add_predictors_old(
  wiz_frame = NULL,
  variable = NULL,
  category = NULL,
  lookback = hours(48),
  window = lookback,
  stats = c(mean = mean, min = min, max = max),
  impute = TRUE,
  output_file = TRUE
)
}
\description{
Old add predictors function using for() loops
}
