% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_predictors.R
\name{wiz_add_predictors_streaming}
\alias{wiz_add_predictors_streaming}
\title{Function to add predictors
before fixed_start.}
\usage{
wiz_add_predictors_streaming(
  wiz_frame = NULL,
  variables = NULL,
  category = NULL,
  lookback = hours(48),
  window = lookback,
  stats = c(mean = mean, min = min, max = max),
  impute = TRUE,
  output_file = TRUE,
  log_file = TRUE,
  check_size_only = FALSE
)
}
\description{
Function to add predictors
before fixed_start.
}
