% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizard.R
\name{wiz_build_temporal_data_dictionary}
\alias{wiz_build_temporal_data_dictionary}
\title{Determine the names and types of all of the temporal data variables.
This function assumes that the temporal data values may be characters if
some variables are categorical. This is an internal function.}
\usage{
wiz_build_temporal_data_dictionary(
  temporal_data,
  temporal_variable,
  temporal_value
)
}
\description{
Determine the names and types of all of the temporal data variables.
This function assumes that the temporal data values may be characters if
some variables are categorical. This is an internal function.
}
