% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_predictors.R
\name{wiz_calc_stats}
\alias{wiz_calc_stats}
\title{Internal function}
\usage{
wiz_calc_stats(
  groups,
  temporal_id,
  temporal_variable,
  temporal_value,
  stats,
  lookback_converted,
  window_converted,
  temporal_data_of_interest,
  total_num_groups,
  pb
)
}
\description{
Internal function
}
