% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_predictors.R
\name{wiz_calc_stats_furrr}
\alias{wiz_calc_stats_furrr}
\title{Internal function
Does not work because furrr cannot find it when it is in plan('multisession') on Windows
So this function has been hard-coded into the add_predictors_function}
\usage{
wiz_calc_stats_furrr(
  groups,
  temporal_id,
  temporal_variable,
  temporal_value,
  stats,
  lookback_converted,
  window_converted,
  temporal_data_of_interest,
  total_num_groups
)
}
\description{
Internal function
Does not work because furrr cannot find it when it is in plan('multisession') on Windows
So this function has been hard-coded into the add_predictors_function
}
