% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{wiz_combine}
\alias{wiz_combine}
\title{New wiz_combine function}
\usage{
wiz_combine(
  wiz_frame,
  ...,
  files = NULL,
  include_files = TRUE,
  wiz_path = TRUE,
  dplyr_join = dplyr::inner_join,
  log_file = TRUE
)
}
\description{
New wiz_combine function
}
