% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiz_viz.R
\name{wiz_viz_lookback}
\alias{wiz_viz_lookback}
\title{A visualization to to evaluate lookback timespan in preparation for building wizard predictors}
\usage{
wiz_viz_lookback(
  wiz_frame,
  categories = NULL,
  variables = NULL,
  grouping = "variable",
  predictor_type = "rolling",
  n_ids = NULL
)
}
\arguments{
\item{wiz_frame}{a wiz_frame object}

\item{categories}{one or more categories separated by |}

\item{variables}{one or more variables separated by |}

\item{grouping}{options are category or variable}

\item{predictor_type}{options are baseline, rolling, or growing}

\item{n_ids}{number of IDs to include in plot}
}
\value{
plot
}
\description{
A visualization to to evaluate lookback timespan in preparation for building wizard predictors
}
\examples{
wiz_viz_lookback(wiz_frame, categories = "labs", predictor_type = "baseline", )
}
