% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiz_viz.R
\name{wiz_viz_stats}
\alias{wiz_viz_stats}
\title{A visualization to to evaluate stats in preparation for building wizard predictors}
\usage{
wiz_viz_stats(
  wiz_frame,
  categories = NULL,
  variables = NULL,
  grouping = "variable",
  predictor_type = "rolling",
  n_ids = NULL,
  n_breaks = 10
)
}
\arguments{
\item{wiz_frame}{a wiz_frame}

\item{categories}{one or more categories}

\item{variables}{one or more variables}

\item{grouping}{options are category or variable}

\item{predictor_type}{options are baseline, rolling, or growing}

\item{n_ids}{number of IDs to include in plot}

\item{n_breaks}{a number of histogram breaks}
}
\value{
plot
}
\description{
A visualization to to evaluate stats in preparation for building wizard predictors
}
\examples{
wiz_viz_stats(wiz_frame, variables = "cr", predictor_type = "growing", n_ids = 2, n_breaks = 10)
}
