% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_predictors.R
\name{wiz_add_predictors}
\alias{wiz_add_predictors}
\title{Add predictors}
\usage{
wiz_add_predictors(
  wiz_frame = NULL,
  variable = NULL,
  category = NULL,
  lookback = hours(48),
  window = hours(12),
  stats = c(mean = mean, min = min, max = max),
  impute = TRUE,
  output_file = TRUE
)
}
\description{
Add predictors
}
