% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_predictors.R
\name{wiz_add_predictors}
\alias{wiz_add_predictors}
\title{New furrr-enabled add_predictors function
#' New add_predictors function using group_modify}
\usage{
wiz_add_predictors(
  wiz_frame = NULL,
  variables = NULL,
  category = NULL,
  lookback = hours(48),
  window = lookback,
  stats = c(mean = mean, min = min, max = max),
  impute = TRUE,
  output_file = TRUE,
  log_file = TRUE,
  check_size_only = FALSE,
  ...
)
}
\description{
New furrr-enabled add_predictors function
#' New add_predictors function using group_modify
}
