% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizard.R
\name{wiz_frame}
\alias{wiz_frame}
\title{Define wizard frame}
\usage{
wiz_frame(
  fixed_data,
  temporal_data,
  fixed_id = "id",
  fixed_start = "",
  fixed_end = "",
  temporal_id = "id",
  temporal_time = "time",
  temporal_variable = "variable",
  temporal_value = "value",
  temporal_category = "variable",
  step = NULL,
  output_folder = NULL
)
}
\description{
Define wizard frame
}
